backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
httpx==0.26.0
python-dotenv==1.0.0
slowapi==0.1.9
supabase==2.3.0
stripe==7.8.0
sentry-sdk[fastapi]==1.39.1

# Shared dependencies from core
Pillow==10.1.0
numpy==1.24.3
aiohttp==3.9.1